+++ api/serverV2.go
@@ -1,6 +1,11 @@
 package api
 
 import (
+	"context"
+
+	"github.com/iotexproject/iotex-election/committee"
+	"google.golang.org/protobuf/proto"
+
 	"github.com/iotexproject/iotex-core/action/protocol"
 	"github.com/iotexproject/iotex-core/actpool"
 	"github.com/iotexproject/iotex-core/blockchain"
@@ -13,8 +18,36 @@ import (
 
 // ServerV2 provides api for user to interact with blockchain data
 type ServerV2 struct {
-	core       *coreService
+	core       *CoreService
 	grpcServer *GrpcServer
+}
+
+// Config represents the config to setup api
+type Config struct {
+	broadcastHandler  BroadcastOutbound
+	electionCommittee committee.Committee
+}
+
+// Option is the option to override the api config
+type Option func(cfg *Config) error
+
+// BroadcastOutbound sends a broadcast message to the whole network
+type BroadcastOutbound func(ctx context.Context, chainID uint32, msg proto.Message) error
+
+// WithBroadcastOutbound is the option to broadcast msg outbound
+func WithBroadcastOutbound(broadcastHandler BroadcastOutbound) Option {
+	return func(cfg *Config) error {
+		cfg.broadcastHandler = broadcastHandler
+		return nil
+	}
+}
+
+// WithNativeElection is the option to return native election data through API.
+func WithNativeElection(committee committee.Committee) Option {
+	return func(cfg *Config) error {
+		cfg.electionCommittee = committee
+		return nil
+	}
 }
 
 // NewServerV2 creates a new server with coreService and GRPC Server
@@ -60,4 +93,8 @@ func (svr *ServerV2) Stop() error {
 		return err
 	}
 	return nil
+}
+
+func (svr *ServerV2) CoreService() *CoreService {
+	return svr.core
 }
